introduce client.endpoint package
